refactor(@molt/command): extract type module
