+++ packages/@molt/command/src/CommandParameter/CommandParameter.ts
@@ -5,7 +5,7 @@ export * from './transform.js'
 export * from './types.js'
 export * from './validate.js'
 import { stripeNegatePrefix } from '../helpers.js'
-import type { Pam } from '../lib/Pam/index.js'
+import type { Type } from '../Type/index.js'
 import type { Output } from './output.js'
 
 export const findByName = (name: string, specs: Output[]): null | Output => {
@@ -55,7 +55,7 @@ export const hasName = (spec: Output, name: string): null | NameHit => {
   return result
 }
 
-export const isOrHasType = (spec: Output, typeTag: Pam.Type['_tag']): boolean => {
+export const isOrHasType = (spec: Output, typeTag: Type.Type['_tag']): boolean => {
   return spec.type._tag === `TypeUnion`
     ? spec.type.members.find((_) => _.type._tag === typeTag) !== undefined
     : spec.type._tag === typeTag
