+++ packages/@molt/command/src/Type/index.ts
@@ -1,0 +1,1 @@
+export * as Type from './Type.js'
