+++ packages/@molt/command/src/Type/types/Scalar.ts
@@ -1,0 +1,5 @@
+import type { Boolean, Enumeration, Literal, Number, String } from './Scalars/index.js'
+
+export type Scalar = Literal | Boolean | Enumeration | String | Number // eslint-disable-line
+
+export * as Scalar from './Scalars/index.js'
