+++ packages/@molt/command/src/Type/types/Scalars/Boolean.ts
@@ -1,0 +1,8 @@
+export interface Boolean {
+  _tag: 'TypeBoolean'
+}
+
+// eslint-disable-next-line
+export const boolean = (): Boolean => {
+  return { _tag: `TypeBoolean` }
+}
