+++ packages/@molt/command/src/Type/types/Scalars/Literal.ts
@@ -1,0 +1,10 @@
+export interface Literal<$Value extends Value = Value> {
+  _tag: 'TypeLiteral'
+  value: $Value
+}
+
+type Value = number | string | boolean
+
+export const literal = <const $Value extends Value>(value: $Value): Literal<$Value> => {
+  return { _tag: `TypeLiteral`, value }
+}
