+++ packages/@molt/command/src/Type/types/Scalars/Number.ts
@@ -1,0 +1,13 @@
+export interface Number {
+  _tag: 'TypeNumber'
+  int?: boolean
+  min?: number
+  max?: number
+  multipleOf?: number
+  finite?: boolean
+}
+
+// eslint-disable-next-line
+export const number = (): Number => {
+  return { _tag: `TypeNumber` }
+}
