+++ packages/@molt/command/src/Type/types/Scalars/String.ts
@@ -1,0 +1,53 @@
+export interface String {
+  _tag: 'TypeString'
+  transformations?: {
+    trim?: boolean
+    toCase?: 'upper' | 'lower'
+  }
+  regex?: RegExp
+  min?: number
+  max?: number
+  length?: number
+  pattern?:
+    | {
+        type: 'email'
+      }
+    | {
+        type: 'url'
+      }
+    | {
+        type: 'uuid'
+      }
+    | {
+        type: 'cuid'
+      }
+    | {
+        type: 'cuid2'
+      }
+    | {
+        type: 'ulid'
+      }
+    | {
+        type: 'emoji'
+      }
+    | {
+        type: 'ip'
+        /**
+         * If `null` then either IPv4 or IPv6 is allowed.
+         */
+        version: 4 | 6 | null
+      }
+    | {
+        type: 'dateTime'
+        offset: boolean
+        precision: null | number
+      }
+  startsWith?: string
+  endsWith?: string
+  includes?: string
+}
+
+// eslint-disable-next-line
+export const string = (): String => {
+  return { _tag: `TypeString` }
+}
