+++ packages/@molt/command/src/TypeAdaptors/TypeAdaptors.ts
@@ -1,0 +1,1 @@
+export * from './zod/index.js'
