+++ packages/@molt/command/src/TypeAdaptors/index.ts
@@ -1,0 +1,1 @@
+export * as TypeAdaptors from './TypeAdaptors.js'
