+++ packages/@molt/command/src/TypeAdaptors/zod/index.ts
@@ -1,0 +1,1 @@
+export * as Zod from './zod.js'
