+++ packages/@molt/command/src/TypeAdaptors/zod/zod.ts
@@ -1,0 +1,10 @@
+import type { Type } from '../../Type/Type.js'
+import type { z } from 'zod'
+
+export * from './helpers.js'
+
+export const fromZod = (type: z.ZodType): Type => {
+  return {
+    _tag: `TypeString`,
+  }
+}
