--- packages/@molt/command/src/lib/Pam/helpers.ts
@@ -1,10 +1,0 @@
-import type { Type } from './kinds.js'
-
-// prettier-ignore
-export type TypeToValueMapping<T extends Type> =
-  T extends Type.Scalar.Boolean         ? boolean :
-  T extends Type.Scalar.Number          ? number  :
-  T extends Type.Scalar.String          ? string  :
-  T extends Type.Scalar.Enumeration     ? string  :
-  T extends Type.Union                  ? boolean | number | string :
-                                          never
