+++ packages/@molt/command/src/lib/Pam/index_.ts
@@ -1,4 +1,2 @@
-export * from './helpers.js'
-export * from './kinds.js'
 export * from './parameter.js'
 export * from './types.js'
