+++ packages/@molt/command/src/lib/Prompter/Input/_core.ts
@@ -1,4 +1,5 @@
 import { casesExhausted } from '../../../helpers.js'
+import type { Type } from '../../../Type/index.js'
 import type { Pam } from '../../Pam/index.js'
 import type { PromptEngine } from '../../PromptEngine/PromptEngine.js'
 import * as Inputs from './types/index.js'
@@ -18,23 +19,23 @@ export const inputForParameter = (params: Params<Pam.Parameter>) => {
   }
 
   if (parameter.type._tag === `TypeUnion`) {
-    return Inputs.union(params as Params<Pam.Parameter<Pam.Type.Union>>)
+    return Inputs.union(params as Params<Pam.Parameter<Type.Union>>)
   }
 
   if (parameter.type._tag === `TypeBoolean`) {
-    return Inputs.boolean(params as Params<Pam.Parameter<Pam.Type.Scalar.Boolean>>)
+    return Inputs.boolean(params as Params<Pam.Parameter<Type.Scalar.Boolean>>)
   }
 
   if (parameter.type._tag === `TypeString`) {
-    return Inputs.string(params as Params<Pam.Parameter<Pam.Type.Scalar.String>>)
+    return Inputs.string(params as Params<Pam.Parameter<Type.Scalar.String>>)
   }
 
   if (parameter.type._tag === `TypeNumber`) {
-    return Inputs.number(params as Params<Pam.Parameter<Pam.Type.Scalar.Number>>)
+    return Inputs.number(params as Params<Pam.Parameter<Type.Scalar.Number>>)
   }
 
   if (parameter.type._tag === `TypeEnum`) {
-    return Inputs.enumeration(params as Params<Pam.Parameter<Pam.Type.Scalar.Enumeration>>)
+    return Inputs.enumeration(params as Params<Pam.Parameter<Type.Scalar.Enumeration>>)
   }
 
   throw casesExhausted(parameter.type)
